+++ NEWS.md
@@ -1,5 +1,8 @@
 # usethis (development version)
 
+* `proj_activate()` now has `new_session`, which controls whether to open a project in
+  a new RStudio session or switch projct (@olivroy, #1954).
+
 * The `ui_*()` functions have been marked as
   [superseded](https://lifecycle.r-lib.org/articles/stages.html#superseded).
   External users of these functions are encouraged to use the